QuantumInfo Documentation

Functions

Missing docstring.

Missing docstring for basis_vector. Check Documenter's build log for details.

QuantumInfo.ket — Function

ket

Returns a vector corresponding to a particular basis unit vector in some Hilbert space.

The vector can be specified in a number of different ways:

by a string and a base

by a integer and a dimension

If the vector is specified by a string and a base, the dimension is inferred from the length of the string. The base is assumed to be 2 unless explicitly specified.

If the basis element is specified by an integer, the dimension is assumed to be 2 unless explicitly specified. The integer values allowed are 0 through the dimension minus 1.

Optionally, the type of the elements of the vector can be specified in the first argument.

QuantumInfo.bra — Function

bra

Returns a dual vector (row vector) corresponding to a particular basis unit vector in some Hilbert space.

See ket for details on how the corresponding vector may be specified.

QuantumInfo.ketbra — Function

ketbra

Returns the outer product between a dual vector (row vector) and a vector, each corresponding to a particular basis unit vector in some Hilbert space. This can also be interpreted as a matrix unite element specified by two unit vectors for a row and a column.

See ket for details on how the corresponding vectors may be specified.

QuantumInfo.projector — Function

projector(v)

Computes the rank-1 projector operator corresponding to a given vector. If more than one vector is given, the projector into the space spanned by the vectors is computed.

QuantumInfo.depol — Function

depol(d, p=1.0)

depol return a superoperator that replaces the input with a maximally mixed state with probability p, and leaves it unchanged with probability (1-p).

QuantumInfo.unitalproj — Function

unitalproj(m)

Given a superoperator j in a Liouville representation, unitalproj extracts the closest superoperator (in Frobenius norm) that is unital. The result may not be completely positive.

Missing docstring.

Missing docstring for ishermitian. Check Documenter's build log for details.

QuantumInfo.iscp — Function

iscp(m; tol)

Checks if the liouville representation of a map (in the natural, computational basis) is completely positive.

QuantumInfo.istp — Function

istp(m; tol)

Checks if the Liouville representation of a map is trace preserving (TP).

QuantumInfo.ischannel — Function

ischannel(m; tol)

Checks if the Liouville representation of a map is completely positive (CP) and trace preserving (TP).

QuantumInfo.isunital — Function

isunital(m; tol)

Checks the conditions for unitality of a map in Liouville representation.

QuantumInfo.isliouvillian — Function

isliouvillian(m; tol)

Checks the conditions for a physical Liouvillian matrix (CPTP map generator)

Missing docstring.

Missing docstring for liouvillian_violation. Check Documenter's build log for details.

QuantumInfo.nearestu — Function

nearestu(l)

Computes the unitary CP map closest to a given CP map in an interferometric sense. See D. Oi, Phys. Rev. Lett. 91, 067902 (2003)

Data Structs

Missing docstring.

Missing docstring for AvgFidelity. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DepolRate. Check Documenter's build log for details.

Index

QuantumInfo.bra

QuantumInfo.depol

QuantumInfo.ischannel

QuantumInfo.iscp

QuantumInfo.isliouvillian

QuantumInfo.istp

QuantumInfo.isunital

QuantumInfo.ket

QuantumInfo.ketbra

QuantumInfo.nearestu

QuantumInfo.projector

QuantumInfo.unitalproj